+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Deprecated]
 ### [Removed]
 ### [Fixed]
+- @aofl/cli - Serve only opens 2 browser tab
+
 ### [Security]
 ---