+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.5] - 2019-06-20
+### [Added]
+- @aofl/cli - debug reporter
+
+### [Changed]
+### [Deprecated]
+### [Removed]
+### [Fixed]
+- @aofl/cli - exits with error code when compiler reports an error
+### [Security]
+
+---
 ## [2.0.4] - 2019-06-18
 ### [Added]
 - @aofl/cli - reporter option for build and test
@@ -664,6 +676,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ---
 
 
+[2.0.4]:https://github.com/AgeOfLearning/aofl/compare/v2.0.4...v2.0.5
 [2.0.4]:https://github.com/AgeOfLearning/aofl/compare/v2.0.3...v2.0.4
 [2.0.3]:https://github.com/AgeOfLearning/aofl/compare/v2.0.2...v2.0.3
 [2.0.2]:https://github.com/AgeOfLearning/aofl/compare/v2.0.1...v2.0.2